requirements.txt
flask==3.1.3
inflect==7.5.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
python-dotenv==1.2.2